Adds a way of totalling sums
This commit adds a couple of methds. The first, _total, takes one of the people as an argument and generates a dictionary. The dictionary has a tuple as a key, containing the indices, and the sum of the integers between them as the value. The second method, _totals, returns a dictionary with a string (the name of the person) as the key and the dictionary returned from _total as the value.